Phase 4e: move firmware/delta resolution from Devices to Firmwares
Continues the Devices cleanup by relocating logic that isn't about devices.
Built on Phase 4d (cleanup/phase-4d-devices-certificates).

Moves the firmware/delta-resolution functions out of NervesHub.Devices into
NervesHub.Firmwares, where the firmware/delta schemas, the update-tool config,
and the firmware getters already live:
- get_delta_or_firmware/2, get_delta_if_ready/3 (private)
- get_delta_url/2, delta_ready?/2, delta_updatable?/2

These take a device only to read its current firmware_metadata; they query
Firmware/FirmwareDelta and call other Firmwares functions, so Firmwares is
their natural home.

Also:
- Deletes the duplicate update_tool/0 from Devices (Firmwares already has an
  identical private copy, now the single source).
- Dedups the source-firmware subquery that delta_ready?/2 and get_delta_url/2
  each built inline into a shared private firmware_delta_query/3.
- Drops the now-unused Fwup and DeploymentRelease aliases from Devices.

Call sites (device_events, details_tab, resolve_update, tests) repointed to
Firmwares.

Verification: mix format, credo, dialyzer (0 unskipped), full test suite
(1396 passing; 1 unrelated ConnectionHistoryTest DBConnection flake).
