+++ lib/nerves_hub_web/components/device_page/details_tab.ex
@@ -778,8 +778,8 @@ defmodule NervesHubWeb.Components.DevicePage.DetailsTab do
 
     {:ok, firmware} = Firmwares.get_firmware_by_product_and_uuid(product, uuid)
 
-    firmware_delta_updatable? = Devices.delta_updatable?(device, firmware)
-    delta_complete? = Devices.delta_ready?(device, firmware)
+    firmware_delta_updatable? = Firmwares.delta_updatable?(device, firmware)
+    delta_complete? = Firmwares.delta_ready?(device, firmware)
 
     socket
     |> assign(:delta_available?, firmware_delta_updatable? && delta_complete?)
